refactor: api 파일에서 타입 분리
